+++ src/main/groovy/com/github/blazsolar/gradle/task/Pmd.groovy
@@ -1,5 +1,8 @@
 package com.github.blazsolar.gradle.task
 
+import com.android.SdkConstants
+import com.android.build.gradle.api.AndroidSourceSet
+import com.github.blazsolar.gradle.extensions.PMDExtension
 import org.gradle.api.tasks.TaskAction
 
 /**
@@ -14,4 +17,24 @@ class Pmd extends org.gradle.api.plugins.quality.Pmd {
 
         super.run()
     }
+
+    void setup(PMDExtension extension) {
+
+        description "PMD for debug source"
+        group "Check"
+
+        ignoreFailures = false
+        ruleSets = ["java-basic", "java-braces", "java-strings", "java-android"]
+
+        project.android.sourceSetsContainer.all { AndroidSourceSet sourceSet ->
+            if (!sourceSet.name.startsWith("test") && !sourceSet.name.startsWith(SdkConstants.FD_TEST)) {
+                source sourceSet.java.srcDirs
+            }
+        }
+
+        include '**/*.java'
+        exclude '**/gen/**'
+
+    }
+
 }